tests

+++ src/main/java/com/technicaltest/services/VehicleService.java
@@ -7,6 +7,7 @@ import com.technicaltest.models.VehicleEntity;
 import com.technicaltest.repositories.VehicleRepository;
 import com.technicaltest.utils.Constants;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -19,6 +20,13 @@ import java.util.*;
 @Service
 public class VehicleService {
 
+    public static final String MODEL = "model";
+    public static final String YEAR = "year";
+    public static final String BRAND_ENTITY_NAME = "brandEntity.name";
+    public static final String LICENSE_PLATE = "licensePlate";
+    public static final String VEHICLE_DOES_NOT_EXIST = "El vehiculo no existe";
+    private static final String VEHICLES = "vehicles";
+    private static final String TOTAL_PAGES = "totalPages";
     private static final String GET_VEHICLES_ERROR = "Error al obtener los vehiculos";
     private final VehicleRepository vehicleRepository;
     private final BrandService brandService;
@@ -30,24 +38,19 @@ public class VehicleService {
     }
 
     public ResponseDTO getVehicles(int page, int size, String sortField, String sortDirection) {
-
-        try {
             validateSortField(sortField);
             Page<VehicleEntity> vehiclePage = orderAndPageVehicles(page, size, sortField, sortDirection);
             Map<String, Object> response = new HashMap<>();
-            response.put("vehicles", vehiclePage.getContent());
-            response.put("totalPages", vehiclePage.getTotalPages() - 1);
+            response.put(VEHICLES, vehiclePage.getContent());
+            response.put(TOTAL_PAGES, vehiclePage.getTotalPages() - 1);
             return ResponseDTO.builder()
                     .response(response)
                     .error(false)
                     .build();
-        } catch (Exception e) {
-            throw new GlobalException(GET_VEHICLES_ERROR);
-        }
     }
 
     private void validateSortField(String sortField) {
-        List<String> validFields = Arrays.asList("model", "year", "brandEntity.name");
+        List<String> validFields = Arrays.asList(MODEL, YEAR, BRAND_ENTITY_NAME, LICENSE_PLATE);
         if (!validFields.contains(sortField)) {
             throw new IllegalArgumentException("Campo invalido para ordenar: " + sortField);
         }
@@ -60,36 +63,35 @@ public class VehicleService {
 
     private PageRequest createPageRequest(int page, int size, String sortField, String sortDirection) {
         Sort sort = Sort.by(sortField);
-        sort = sortDirection.equalsIgnoreCase("desc") ? sort.ascending() : sort.descending();
+        sort = sortDirection.equalsIgnoreCase("asc") ? sort.ascending() : sort.descending();
         return PageRequest.of(page, size, sort);
     }
 
     public ResponseDTO searchVehicles(String search, int page, int size) {
-        try {
 
             Pageable pageable = PageRequest.of(page, size);
             Page<VehicleEntity> vehiclePage = vehicleRepository.findByBrandModelOrLicensePlate(search, pageable);
             Map<String, Object> response = new HashMap<>();
-            response.put("vehicles", vehiclePage.getContent());
-            response.put("totalPages", vehiclePage.getTotalPages() - 1 );
+            response.put(VEHICLES, vehiclePage.getContent());
+            response.put(TOTAL_PAGES, vehiclePage.getTotalPages() - 1 );
 
             return ResponseDTO.builder()
                     .response(response)
                     .error(false)
                     .build();
-        } catch (Exception e) {
-            throw new GlobalException(GET_VEHICLES_ERROR);
-        }
     }
 
     public ResponseDTO getVehicleById(String id) {
         return ResponseDTO.builder()
-                .response(vehicleRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("El vehiculo no existe")))
+                .response(vehicleRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(VEHICLE_DOES_NOT_EXIST)))
                 .error(false)
                 .build();
     }
 
     public ResponseDTO addVehicle(VehicleDTO vehicleDTO) {
+        if(vehicleRepository.existsByLicensePlate(vehicleDTO.getLicensePlate())){
+            throw new DataIntegrityViolationException("Ya existe un vehiculo con esa placa");
+        }
 
         VehicleEntity vehicleEntity = VehicleEntity.builder()
                 .id(UUID.randomUUID().toString())
@@ -97,7 +99,7 @@ public class VehicleService {
                 .licensePlate(vehicleDTO.getLicensePlate())
                 .model(vehicleDTO.getModel())
                 .year(vehicleDTO.getYear())
-                .brandEntity(brandService.getBrandById(vehicleDTO.getBrandId()))
+                .brandEntity(brandService.findBrandById(vehicleDTO.getBrandId()))
                 .build();
 
         vehicleRepository.save((vehicleEntity));
@@ -109,13 +111,17 @@ public class VehicleService {
     }
 
     public ResponseDTO updateVehicle(String id, VehicleDTO vehicleDTO) {
-        VehicleEntity vehicleEntity = vehicleRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("El vehiculo no existe"));
+        VehicleEntity vehicleEntity = vehicleRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(VEHICLE_DOES_NOT_EXIST));
+
+        if(vehicleRepository.existsByLicensePlate(vehicleDTO.getLicensePlate()) && !vehicleEntity.getLicensePlate().equals(vehicleDTO.getLicensePlate())){
+            throw new DataIntegrityViolationException("Ya existe un vehiculo con esa placa");
+        }
 
         vehicleEntity.setColor(vehicleDTO.getColor());
         vehicleEntity.setLicensePlate(vehicleDTO.getLicensePlate());
         vehicleEntity.setModel(vehicleDTO.getModel());
         vehicleEntity.setYear(vehicleDTO.getYear());
-        vehicleEntity.setBrandEntity(brandService.getBrandById(vehicleDTO.getBrandId()));
+        vehicleEntity.setBrandEntity(brandService.findBrandById(vehicleDTO.getBrandId()));
 
         vehicleRepository.save(vehicleEntity);
 
@@ -124,17 +130,15 @@ public class VehicleService {
                 .error(false)
                 .build();
     }
-
     public ResponseDTO deleteVehicle(String id) {
-        try {
+            if(!vehicleRepository.existsById(id)){
+                throw new EntityNotFoundException(VEHICLE_DOES_NOT_EXIST);
+            }
             vehicleRepository.deleteById(id);
             return ResponseDTO.builder()
                     .response(Constants.SUCCESS)
                     .error(false)
                     .build();
-        } catch (Exception e) {
-            throw new GlobalException("Error al eliminar el vehiculo");
-        }
     }
 
 }
